+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/assembly/KafkaAssemblyOperator.java
@@ -51,6 +51,7 @@ import org.apache.logging.log4j.Logger;
 
 import java.io.File;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.SortedMap;
 import java.util.TreeMap;
@@ -127,9 +128,12 @@ public class KafkaAssemblyOperator extends AbstractAssemblyOperator<KubernetesCl
                 .compose(state -> state.kafkaScaleDown())
                 .compose(state -> state.kafkaService())
                 .compose(state -> state.kafkaHeadlessService())
+                .compose(state -> state.kafkaExternalBootstrapService())
                 .compose(state -> state.kafkaReplicaServices())
                 .compose(state -> state.kafkaBootstrapRoute())
                 .compose(state -> state.kafkaReplicaRoutes())
+                .compose(state -> state.kafkaExternalBootstrapServiceReady())
+                .compose(state -> state.kafkaReplicaServicesReady())
                 .compose(state -> state.kafkaBootstrapRouteReady())
                 .compose(state -> state.kafkaReplicaRoutesReady())
                 .compose(state -> state.kafkaGenerateCertificates())
@@ -449,6 +453,10 @@ public class KafkaAssemblyOperator extends AbstractAssemblyOperator<KubernetesCl
             return withVoid(serviceOperations.reconcile(namespace, kafkaCluster.getHeadlessServiceName(), kafkaHeadlessService));
         }
 
+        Future<ReconciliationState> kafkaExternalBootstrapService() {
+            return withVoid(serviceOperations.reconcile(namespace, KafkaCluster.externalBootstrapServiceName(name), kafkaCluster.generateExternalBootstrapService()));
+        }
+
         Future<ReconciliationState> kafkaReplicaServices() {
             int replicas = kafkaCluster.getReplicas();
             List<Future> serviceFutures = new ArrayList<>(replicas);
@@ -491,32 +499,166 @@ public class KafkaAssemblyOperator extends AbstractAssemblyOperator<KubernetesCl
             return withVoid(CompositeFuture.join(routeFutures));
         }
 
+        Future<ReconciliationState> kafkaExternalBootstrapServiceReady() {
+            if (!kafkaCluster.isExposedWithLoadBalancer() && !kafkaCluster.isExposedWithNodePort()) {
+                return withVoid(Future.succeededFuture());
+            }
+
+            Future blockingFuture = Future.future();
+
+            vertx.createSharedWorkerExecutor("kubernetes-ops-pool").executeBlocking(
+                future -> {
+                    String serviceName = KafkaCluster.externalBootstrapServiceName(name);
+                    Future<Void> address = null;
+
+                    if (kafkaCluster.isExposedWithNodePort()) {
+                        address = serviceOperations.hasNodePort(namespace, serviceName, 1_000, operationTimeoutMs);
+                    } else {
+                        address = serviceOperations.hasIngressAddress(namespace, serviceName, 1_000, operationTimeoutMs);
+                    }
+
+                    address.setHandler(res -> {
+                        if (res.succeeded()) {
+                            if (kafkaCluster.isExposedWithLoadBalancer()) {
+                                if (serviceOperations.get(namespace, serviceName).getStatus().getLoadBalancer().getIngress().get(0).getHostname() != null) {
+                                    this.kafkaExternalBootstrapAddress = serviceOperations.get(namespace, serviceName).getStatus().getLoadBalancer().getIngress().get(0).getHostname();
+                                } else {
+                                    this.kafkaExternalBootstrapAddress = serviceOperations.get(namespace, serviceName).getStatus().getLoadBalancer().getIngress().get(0).getIp();
+                                }
+                            } else if (kafkaCluster.isExposedWithNodePort()) {
+                                this.kafkaExternalBootstrapAddress = serviceOperations.get(namespace, serviceName).getSpec().getPorts().get(0).getNodePort().toString();
+                            }
+
+                            if (log.isTraceEnabled()) {
+                                log.trace("{}: Found address {} for Service {}", reconciliation, this.kafkaExternalBootstrapAddress, serviceName);
+                            }
+
+                            future.complete();
+                        } else {
+                            log.warn("{}: No address found for Service {}", reconciliation, serviceName);
+                            future.fail("No address found for Service " + serviceName);
+                        }
+                    });
+                }, res -> {
+                    if (res.succeeded())    {
+                        blockingFuture.complete();
+                    } else  {
+                        blockingFuture.fail(res.cause());
+                    }
+                });
+
+            return withVoid(blockingFuture);
+        }
+
+        Future<ReconciliationState> kafkaReplicaServicesReady() {
+            if (!kafkaCluster.isExposedWithLoadBalancer() && !kafkaCluster.isExposedWithNodePort()) {
+                return withVoid(Future.succeededFuture());
+            }
+
+            Future blockingFuture = Future.future();
+
+            vertx.createSharedWorkerExecutor("kubernetes-ops-pool").executeBlocking(
+                future -> {
+                    int replicas = kafkaCluster.getReplicas();
+                    List<Future> routeFutures = new ArrayList<>(replicas);
+
+                    for (int i = 0; i < replicas; i++) {
+                        String serviceName = KafkaCluster.externalServiceName(name, i);
+                        Future routeFuture = Future.future();
+
+                        Future<Void> address = null;
+
+                        if (kafkaCluster.isExposedWithNodePort()) {
+                            address = serviceOperations.hasNodePort(namespace, serviceName, 1_000, operationTimeoutMs);
+                        } else {
+                            address = serviceOperations.hasIngressAddress(namespace, serviceName, 1_000, operationTimeoutMs);
+                        }
+
+                        int podNumber = i;
+
+                        address.setHandler(res -> {
+                            if (res.succeeded()) {
+                                String serviceAddress = null;
+                                if (kafkaCluster.isExposedWithLoadBalancer()) {
+                                    if (serviceOperations.get(namespace, serviceName).getStatus().getLoadBalancer().getIngress().get(0).getHostname() != null) {
+                                        serviceAddress = serviceOperations.get(namespace, serviceName).getStatus().getLoadBalancer().getIngress().get(0).getHostname();
+                                    } else {
+                                        serviceAddress = serviceOperations.get(namespace, serviceName).getStatus().getLoadBalancer().getIngress().get(0).getIp();
+                                    }
+                                } else if (kafkaCluster.isExposedWithNodePort()) {
+                                    serviceAddress = serviceOperations.get(namespace, serviceName).getSpec().getPorts().get(0).getNodePort().toString();
+                                }
+
+                                this.kafkaExternalAddresses.put(podNumber, serviceAddress);
+
+                                if (log.isTraceEnabled()) {
+                                    log.trace("{}: Found address {} for Service {}", reconciliation, serviceAddress, serviceName);
+                                }
+
+                                routeFuture.complete();
+                            } else {
+                                log.warn("{}: No address found for Service {}", reconciliation, serviceName);
+                                routeFuture.fail("No address found for Service " + serviceName);
+                            }
+                        });
+
+                        routeFutures.add(routeFuture);
+                    }
+
+                    CompositeFuture.join(routeFutures).setHandler(res -> {
+                        if (res.succeeded()) {
+                            future.complete();
+                        } else  {
+                            future.fail(res.cause());
+                        }
+                    });
+                }, res -> {
+                    if (res.succeeded())    {
+                        blockingFuture.complete();
+                    } else  {
+                        blockingFuture.fail(res.cause());
+                    }
+                });
+
+            return withVoid(blockingFuture);
+        }
 
         Future<ReconciliationState> kafkaBootstrapRouteReady() {
             if (routeOperations == null || !kafkaCluster.isExposedWithRoute()) {
                 return withVoid(Future.succeededFuture());
             }
 
-            String routeName = KafkaCluster.serviceName(name);
-            Future future = Future.future();
-            Future<Void> address = routeOperations.hasAddress(namespace, routeName, 1_000, operationTimeoutMs);
+            Future blockingFuture = Future.future();
 
-            address.setHandler(res -> {
-                if (res.succeeded())    {
-                    this.kafkaExternalBootstrapAddress = routeOperations.get(namespace, routeName).getSpec().getHost();
+            vertx.createSharedWorkerExecutor("kubernetes-ops-pool").executeBlocking(
+                future -> {
+                    String routeName = KafkaCluster.serviceName(name);
+                    //Future future = Future.future();
+                    Future<Void> address = routeOperations.hasAddress(namespace, routeName, 1_000, operationTimeoutMs);
 
-                    if (log.isTraceEnabled()) {
-                        log.trace("{}: Found address {} for Route {}", reconciliation, routeOperations.get(namespace, routeName).getSpec().getHost(), routeName);
+                    address.setHandler(res -> {
+                        if (res.succeeded()) {
+                            this.kafkaExternalBootstrapAddress = routeOperations.get(namespace, routeName).getSpec().getHost();
+
+                            if (log.isTraceEnabled()) {
+                                log.trace("{}: Found address {} for Route {}", reconciliation, routeOperations.get(namespace, routeName).getSpec().getHost(), routeName);
+                            }
+
+                            future.complete();
+                        } else {
+                            log.warn("{}: No address found for Route {}", reconciliation, routeName);
+                            future.fail("No address found for Route " + routeName);
+                        }
+                    });
+                }, res -> {
+                    if (res.succeeded())    {
+                        blockingFuture.complete();
+                    } else  {
+                        blockingFuture.fail(res.cause());
                     }
+                });
 
-                    future.complete();
-                } else {
-                    log.warn("{}: No address found for Route {}", reconciliation, routeName);
-                    future.fail("No address found for Route " + routeName);
-                }
-            });
-
-            return withVoid(future);
+            return withVoid(blockingFuture);
         }
 
         Future<ReconciliationState> kafkaReplicaRoutesReady() {
@@ -524,38 +666,62 @@ public class KafkaAssemblyOperator extends AbstractAssemblyOperator<KubernetesCl
                 return withVoid(Future.succeededFuture());
             }
 
-            int replicas = kafkaCluster.getReplicas();
-            List<Future> routeFutures = new ArrayList<>(replicas);
+            Future blockingFuture = Future.future();
 
-            for (int i = 0; i < replicas; i++) {
-                String routeName = KafkaCluster.externalServiceName(name, i);
-                Future future = Future.future();
-                Future<Void> address = routeOperations.hasAddress(namespace, routeName, 1_000, operationTimeoutMs);
-                int podNumber = i;
+            vertx.createSharedWorkerExecutor("kubernetes-ops-pool").executeBlocking(
+                future -> {
+                    int replicas = kafkaCluster.getReplicas();
+                    List<Future> routeFutures = new ArrayList<>(replicas);
 
-                address.setHandler(res -> {
-                    if (res.succeeded()) {
-                        this.kafkaExternalAddresses.put(podNumber, routeOperations.get(namespace, routeName).getSpec().getHost());
+                    for (int i = 0; i < replicas; i++) {
+                        String routeName = KafkaCluster.externalServiceName(name, i);
+                        Future routeFuture = Future.future();
+                        Future<Void> address = routeOperations.hasAddress(namespace, routeName, 1_000, operationTimeoutMs);
+                        int podNumber = i;
 
-                        if (log.isTraceEnabled()) {
-                            log.trace("{}: Found address {} for Route {}", reconciliation, routeOperations.get(namespace, routeName).getSpec().getHost(), routeName);
+                        address.setHandler(res -> {
+                            if (res.succeeded()) {
+                                this.kafkaExternalAddresses.put(podNumber, routeOperations.get(namespace, routeName).getSpec().getHost());
+
+                                if (log.isTraceEnabled()) {
+                                    log.trace("{}: Found address {} for Route {}", reconciliation, routeOperations.get(namespace, routeName).getSpec().getHost(), routeName);
+                                }
+
+                                routeFuture.complete();
+                            } else {
+                                log.warn("{}: No address found for Route {}", reconciliation, routeName);
+                                routeFuture.fail("No address found for Route " + routeName);
+                            }
+                        });
+
+                        routeFutures.add(routeFuture);
+                    }
+
+                    CompositeFuture.join(routeFutures).setHandler(res -> {
+                        if (res.succeeded()) {
+                            future.complete();
+                        } else  {
+                            future.fail(res.cause());
                         }
-
-                        future.complete();
-                    } else {
-                        log.warn("{}: No address found for Route {}", reconciliation, routeName);
-                        future.fail("No address found for Route " + routeName);
+                    });
+                }, res -> {
+                    if (res.succeeded())    {
+                        blockingFuture.complete();
+                    } else  {
+                        blockingFuture.fail(res.cause());
                     }
                 });
 
-                routeFutures.add(future);
-            }
-
-            return withVoid(CompositeFuture.join(routeFutures));
+            return withVoid(blockingFuture);
         }
 
         Future<ReconciliationState> kafkaGenerateCertificates() {
-            kafkaCluster.generateCertificates(certManager, assemblySecrets, kafkaExternalBootstrapAddress, kafkaExternalAddresses);
+            if (kafkaCluster.isExposedWithNodePort()) {
+                kafkaCluster.generateCertificates(certManager, assemblySecrets, null, Collections.EMPTY_MAP);
+            } else  {
+                kafkaCluster.generateCertificates(certManager, assemblySecrets, kafkaExternalBootstrapAddress, kafkaExternalAddresses);
+            }
+
             return withVoid(Future.succeededFuture());
         }