+++ operator-common/src/test/java/io/strimzi/operator/common/operator/resource/ServiceOperatorTest.java
@@ -5,14 +5,20 @@
 package io.strimzi.operator.common.operator.resource;
 
 import io.fabric8.kubernetes.api.model.DoneableService;
+import io.fabric8.kubernetes.api.model.IntOrString;
 import io.fabric8.kubernetes.api.model.Service;
 import io.fabric8.kubernetes.api.model.ServiceBuilder;
 import io.fabric8.kubernetes.api.model.ServiceList;
+import io.fabric8.kubernetes.api.model.ServicePortBuilder;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.dsl.MixedOperation;
 import io.fabric8.kubernetes.client.dsl.Resource;
 import io.vertx.core.Vertx;
+import io.vertx.ext.unit.TestContext;
+import org.junit.Test;
 
+import static org.junit.Assert.assertEquals;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 public class ServiceOperatorTest extends AbstractResourceOperatorTest<KubernetesClient, Service, ServiceList, DoneableService, Resource<Service, DoneableService>> {
@@ -40,5 +46,62 @@ public class ServiceOperatorTest extends AbstractResourceOperatorTest<Kubernetes
     @Override
     protected ServiceOperator createResourceOperations(Vertx vertx, KubernetesClient mockClient) {
         return new ServiceOperator(vertx, mockClient);
+    }
+
+    @Test
+    public void testNodePortPatching(TestContext context)  {
+        KubernetesClient client = mock(KubernetesClient.class);
+
+        Service current = new ServiceBuilder()
+                .withNewMetadata()
+                    .withNamespace(NAMESPACE)
+                    .withName(RESOURCE_NAME)
+                .endMetadata()
+                .withNewSpec()
+                    .withType("NodePort")
+                    .withPorts(
+                            new ServicePortBuilder()
+                                    .withName("port1")
+                                    .withPort(1234)
+                                    .withTargetPort(new IntOrString(1234))
+                                    .withNodePort(31234)
+                                    .build(),
+                            new ServicePortBuilder()
+                                    .withName("port2")
+                                    .withPort(5678)
+                                    .withTargetPort(new IntOrString(5678))
+                                    .withNodePort(35678)
+                                    .build()
+                    )
+                .endSpec()
+                .build();
+
+        Service desired = new ServiceBuilder()
+                .withNewMetadata()
+                .withNamespace(NAMESPACE)
+                .withName(RESOURCE_NAME)
+                .endMetadata()
+                .withNewSpec()
+                .withType("NodePort")
+                .withPorts(
+                        new ServicePortBuilder()
+                                .withName("port2")
+                                .withPort(5678)
+                                .withTargetPort(new IntOrString(5678))
+                                .build(),
+                        new ServicePortBuilder()
+                                .withName("port1")
+                                .withPort(1234)
+                                .withTargetPort(new IntOrString(1234))
+                                .build()
+                )
+                .endSpec()
+                .build();
+
+        ServiceOperator op = new ServiceOperator(vertx, client);
+        op.patchNodePorts(current, desired);
+
+        assertEquals(current.getSpec().getPorts().get(0).getNodePort(), desired.getSpec().getPorts().get(1).getNodePort());
+        assertEquals(current.getSpec().getPorts().get(1).getNodePort(), desired.getSpec().getPorts().get(0).getNodePort());
     }
 }