+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## 0.7.0
 
-* Exposing Kafka to the outside of OpenShift using OpenShift Routes
+* Exposing Kafka to the outside using:
+  * OpenShift Routes
+  * LoadBalancers
+  * NodePorts
 * Use less wide RBAC permissions (`ClusterRoleBindings` where converted to `RoleBindings` where possible)
 * Support for SASL authentication using the SCRAM-SHA-512 mechanism added to Kafka Connect and Kafka Connect with S2I support 
 * Network policies for managing access to Zookeeper ports and Kafka replication ports